graph, server: Move explorer::Cache to graph::util::TimedCache
