+++ graph/src/util/mod.rs
@@ -9,6 +9,8 @@ pub mod security;
 
 pub mod lfu_cache;
 
+pub mod timed_cache;
+
 pub mod error;
 
 pub mod stats;
