+++ server/index-node/src/explorer.rs
@@ -8,10 +8,9 @@ use hyper::header::{
 };
 use hyper::Body;
 use std::{
-    collections::HashMap,
     env,
     str::FromStr,
-    sync::{Arc, RwLock},
+    sync::Arc,
     time::{Duration, Instant},
 };
 
@@ -23,6 +22,7 @@ use graph::{
     data::subgraph::status,
     object,
     prelude::{lazy_static, q, serde_json, warn, Logger, SerializableValue},
+    util::timed_cache::TimedCache,
 };
 
 lazy_static! {
@@ -75,9 +75,9 @@ lazy_static! {
 #[derive(Debug)]
 pub struct Explorer<S> {
     store: Arc<S>,
-    versions: Cache<q::Value>,
-    version_infos: Cache<VersionInfo>,
-    entity_counts: Cache<q::Value>,
+    versions: TimedCache<q::Value>,
+    version_infos: TimedCache<VersionInfo>,
+    entity_counts: TimedCache<q::Value>,
 }
 
 impl<S> Explorer<S>
@@ -87,9 +87,9 @@ where
     pub fn new(store: Arc<S>) -> Self {
         Self {
             store,
-            versions: Cache::new(*TTL),
-            version_infos: Cache::new(*TTL),
-            entity_counts: Cache::new(*TTL),
+            versions: TimedCache::new(*TTL),
+            version_infos: TimedCache::new(*TTL),
+            entity_counts: TimedCache::new(*TTL),
         }
     }
 
@@ -255,69 +255,4 @@ fn as_http_response(value: &q::Value) -> http::Response<Body> {
         .header(CONTENT_TYPE, "application/json")
         .body(Body::from(json))
         .unwrap()
-}
-
-/// Caching of values for a specified amount of time
-#[derive(Debug)]
-struct CacheEntry<T> {
-    value: Arc<T>,
-    expires: Instant,
-}
-
-/// A cache that keeps entries live for a fixed amount of time. It is assumed
-/// that all that data that could possibly wind up in the cache is very small,
-/// and that expired entries are replaced by an updated entry whenever expiry
-/// is detected. In other words, the cache does not ever remove entries.
-#[derive(Debug)]
-struct Cache<T> {
-    ttl: Duration,
-    entries: RwLock<HashMap<String, CacheEntry<T>>>,
-}
-
-impl<T> Cache<T> {
-    fn new(ttl: Duration) -> Self {
-        Self {
-            ttl,
-            entries: RwLock::new(HashMap::new()),
-        }
-    }
-
-    /// Return the entry for `key` if it exists and is not expired yet, and
-    /// return `None` otherwise. Note that expired entries stay in the cache
-    /// as it is assumed that, after returning `None`, the caller will
-    /// immediately overwrite that entry with a call to `set`
-    fn get(&self, key: &str) -> Option<Arc<T>> {
-        self.get_at(key, Instant::now())
-    }
-
-    fn get_at(&self, key: &str, now: Instant) -> Option<Arc<T>> {
-        match self.entries.read().unwrap().get(key) {
-            Some(CacheEntry { value, expires }) if *expires >= now => Some(value.clone()),
-            _ => None,
-        }
-    }
-
-    /// Associate `key` with `value` in the cache. The `value` will be
-    /// valid for `self.ttl` duration
-    fn set(&self, key: String, value: Arc<T>) {
-        self.set_at(key, value, Instant::now())
-    }
-
-    fn set_at(&self, key: String, value: Arc<T>, now: Instant) {
-        let entry = CacheEntry {
-            value,
-            expires: now + self.ttl,
-        };
-        self.entries.write().unwrap().insert(key, entry);
-    }
-}
-
-#[test]
-fn cache() {
-    const KEY: &str = "one";
-    let cache = Cache::<String>::new(Duration::from_millis(10));
-    let now = Instant::now();
-    cache.set_at(KEY.to_string(), Arc::new("value".to_string()), now);
-    assert!(cache.get_at(KEY, now + Duration::from_millis(5)).is_some());
-    assert!(cache.get_at(KEY, now + Duration::from_millis(15)).is_none());
 }
